Add firewall opt renewable_auth
which causes an existing authentic security token to be destroyed and
authentication to proceed anew when check_path is requested and there
is a JWT in the request.  Using this option avoids the need to have some
means for the user to logout before they supply a different JWT to the
authentication entry point.